pyproject.toml
[project]
name = "krules-framework"
version = "2.0.0"
description = "KRules Framework - Async event-driven framework"
authors = [{name = "Airspot", email = "info@airspot.tech"}]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["krules", "framework", "event-driven", "async", "reactive"]
homepage = "https://airspot.tech"
repository = "https://github.com/airspot-dev/krules-framework.git"

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "dependency-injector==4.48.2",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
]

[project.optional-dependencies]
pubsub = [
    "google-cloud-pubsub>=2.26.1",
    "cloudevents==1.11.0",
]
fastapi = [
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
]
redis=[
    "redis>=5.2.0",
]
postgres=[
    "asyncpg>=0.30.0",
]
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-localserver>=0.9.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.2",

    # Code Quality
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=7.0.0",
    "pylint>=3.0.0",

    # Type Checking
    "types-pytz",
    "types-toml",

    # Release Management
    "invoke>=2.2.0",
    "semver>=3.0.0",
    "toml>=0.10.2",
    "gitpython>=3.1.40",
    "build>=1.0.3",
    "twine>=4.0.2",

    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[dependency-groups]
test = [
    "pytest>=7.0.0",
    "pytest-localserver>=0.9.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.2",
    "anyio>=4.11.0",
    "pytest-asyncio>=1.2.0",
    # PubSub dependencies for integration tests
    "google-cloud-pubsub>=2.26.1",
    "cloudevents==1.11.0",
    # PostgreSQL dependencies for postgres storage tests
    "asyncpg>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "krules_core",
    "krules_cloudevents",
    "krules_pubsub",
    "krules_cloudevents_pubsub",
    "krules_fastapi_env",
    "redis_subjects_storage",
    "postgres_subjects_storage",
]

[tool.hatch.build]
include = [
    "krules_core/**/*.py",
    "krules_cloudevents/**/*.py",
    "krules_pubsub/**/*.py",
    "krules_cloudevents_pubsub/**/*.py",
    "krules_fastapi_env/**/*.py",
    "redis_subjects_storage/**/*.py",
    "postgres_subjects_storage/**/*.py",
]
exclude = [
    "tests/**/*.py",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
testpaths = [
    "tests/test_core_v2",
    "tests/test_subject",
    "tests/test_cloudevents",
    "tests/test_cloudevents_pubsub",
    "tests/test_redis_subjects_storage",
    "tests/test_postgres_subjects_storage",
    "tests/test_krules_fastapi_env"
]
addopts = "-v"
